requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.0.7
django-crispy-forms==2.3
sqlparse==0.5.1
typing_extensions==4.12.2
